Ad Review comments
Search for lines starting with "# ReviewTS" in the code to find my
comments.

Additional remarks:

I strongly suggest using static analysis tools from an early stage, as
these are invaluable tools for catching bugs that otherwise would only
surface as soon as the code is executed. My recommendations in this regard:
* Add Typehints wherever possible, especially to document parameter and
  return types of user-facing functions. Besides enabling static analysis
  with MyPy, they also serve as valuable documentation for library users.
  See https://docs.python.org/3/library/typing.html for details.
* Apply the following tools to the codebase and strive for a state where
  they do not output any issues. Ideally, one would setup a Github Action
  that runs these every time a commit is pushed:
  * https://pypi.org/project/pylint/ - Linter that checks for actual
    code errors as well as coding style inconsistencies (although the
    latter can be disabled if you don't care about those)
  * https://mypy.readthedocs.io/en/stable/ - Verify that Typehints match
    actual data types
While adapting an existing codebase for these tools takes some upfront
work, in a relatively small project like this the overhead is not that
high. What you gain is a powerful tool to increase code quality and
reduce the occurence of bugs, which I think is especially important when
providing a library for 3rd party users.